Crate ab_contracts_common

Modules

env

metadata

method

Structs

CustomContractErrorCode

ExitCode

Enums

ContractError

Constants

HOST_CALL_FN

Function name used to make calls from guest to host

HOST_CALL_FN_IMPORT

Import function name used to make calls from guest to host

MAX_CODE_SIZE

Max allowed size of the contract code

MAX_TOTAL_METHOD_ARGS

Max number of arguments in a method.

METADATA_STATIC_NAME_PREFIX

The name of the static variable that contains contract metadata

Traits

Contract

A trait that indicates the struct is a contact.

ContractTrait

A trait that indicates the implementation of a contract trait by a contract.

ContractTraitDefinition

A trait that is implemented for dyn ContractTrait and includes constants related to trait definition.