ab_contracts_common
Enum ContractMethodMetadata

#[repr(u8)]
pub enum ContractMethodMetadata {

    Init1 = 0,
    Init2 = 1,
    Init3 = 2,
    Init4 = 3,
    Init5 = 4,
    Init6 = 5,
    Init7 = 6,
    Init8 = 7,
    Init9 = 8,
    Init10 = 9,
    CallStateless1 = 10,
    CallStateless2 = 11,
    CallStateless3 = 12,
    CallStateless4 = 13,
    CallStateless5 = 14,
    CallStateless6 = 15,
    CallStateless7 = 16,
    CallStateless8 = 17,
    CallStateless9 = 18,
    CallStateless10 = 19,
    CallStatefulRo1 = 20,
    CallStatefulRo2 = 21,
    CallStatefulRo3 = 22,
    CallStatefulRo4 = 23,
    CallStatefulRo5 = 24,
    CallStatefulRo6 = 25,
    CallStatefulRo7 = 26,
    CallStatefulRo8 = 27,
    CallStatefulRo9 = 28,
    CallStatefulRo10 = 29,
    CallStatefulRw1 = 30,
    CallStatefulRw2 = 31,
    CallStatefulRw3 = 32,
    CallStatefulRw4 = 33,
    CallStatefulRw5 = 34,
    CallStatefulRw6 = 35,
    CallStatefulRw7 = 36,
    CallStatefulRw8 = 37,
    CallStatefulRw9 = 38,
    CallStatefulRw10 = 39,
    ViewStateless1 = 40,
    ViewStateless2 = 41,
    ViewStateless3 = 42,
    ViewStateless4 = 43,
    ViewStateless5 = 44,
    ViewStateless6 = 45,
    ViewStateless7 = 46,
    ViewStateless8 = 47,
    ViewStateless9 = 48,
    ViewStateless10 = 49,
    ViewStatefulRo1 = 50,
    ViewStatefulRo2 = 51,
    ViewStatefulRo3 = 52,
    ViewStatefulRo4 = 53,
    ViewStatefulRo5 = 54,
    ViewStatefulRo6 = 55,
    ViewStatefulRo7 = 56,
    ViewStatefulRo8 = 57,
    ViewStatefulRo9 = 58,
    ViewStatefulRo10 = 59,
    EnvRo = 60,
    EnvRw = 61,
    SlotWithAddressRo = 62,
    SlotWithAddressRw = 63,
    SlotWithoutAddressRo = 64,
    SlotWithoutAddressRw = 65,
    Input = 66,
    Output = 67,
    Result = 68,

}

Metadata for smart contact methods.

Metadata encoding consists of this enum variant treated as u8 followed by optional metadata encoding rules specific to metadata type variant (see variant’s description).

This metadata is sufficient to fully reconstruct hierarchy of the type in order to generate language bindings, auto-generate UI forms, etc.

Variants

Init1 = 0

#[init] method with 1 argument.

Initializers are encoded af follows:

Length of method name in bytes (u8)

Method name as UTF-8 bytes

Each argument is encoded as follows:

Argument type as u8, one of:

Self::EnvRo

Self::EnvRw

Self::SlotWithAddressRo

Self::SlotWithAddressRw

Self::SlotWithoutAddressRo

Self::SlotWithoutAddressRw

Self::Input

Self::Output

Self::Result

Length of the argument name in bytes (u8)

Argument name as UTF-8 bytes

Recursive metadata of argument’s type as described in IoTypeMetadata.

NOTE: Result, regardless of whether it is a return type or explicit #[result] argument is encoded as a separate argument and counts towards number of arguments. At the same time self doesn’t count towards the number of arguments as it is implicitly defined by the variant of this struct.

Init2 = 1

#[init] method with 2 arguments.

Encoding is the same as Self::Init1

Init3 = 2

#[init] method with 3 arguments.

Encoding is the same as Self::Init1

Init4 = 3

#[init] method with 4 arguments.

Encoding is the same as Self::Init1

Init5 = 4

#[init] method with 5 arguments.

Encoding is the same as Self::Init1

Init6 = 5

#[init] method with 6 arguments.

Encoding is the same as Self::Init1

Init7 = 6

#[init] method with 7 arguments.

Encoding is the same as Self::Init1

Init8 = 7

#[init] method with 8 arguments.

Encoding is the same as Self::Init1

Init9 = 8

#[init] method with 9 arguments.

Encoding is the same as Self::Init1

Init10 = 9

#[init] method with 10 arguments.

Encoding is the same as Self::Init1

CallStateless1 = 10

Stateless #[update] method with 1 argument (doesn’t have self in its arguments).

Encoding is the same as Self::Init1

CallStateless2 = 11

Stateless #[update] method with 2 arguments (doesn’t have self in its arguments).

Encoding is the same as Self::Init1

CallStateless3 = 12

Stateless #[update] method with 3 arguments (doesn’t have self in its arguments).

Encoding is the same as Self::Init1

CallStateless4 = 13

Stateless #[update] method with 4 arguments (doesn’t have self in its arguments).

Encoding is the same as Self::Init1

CallStateless5 = 14

Stateless #[update] method with 5 arguments (doesn’t have self in its arguments).

Encoding is the same as Self::Init1

CallStateless6 = 15

Stateless #[update] method with 6 arguments (doesn’t have self in its arguments).

Encoding is the same as Self::Init1

CallStateless7 = 16

Stateless #[update] method with 7 arguments (doesn’t have self in its arguments).

Encoding is the same as Self::Init1

CallStateless8 = 17

Stateless #[update] method with 8 arguments (doesn’t have self in its arguments).

Encoding is the same as Self::Init1

CallStateless9 = 18

Stateless #[update] method with 9 arguments (doesn’t have self in its arguments).

Encoding is the same as Self::Init1

CallStateless10 = 19

Stateless #[update] method with 10 arguments (doesn’t have self in its arguments).

Encoding is the same as Self::Init1

CallStatefulRo1 = 20

Stateful read-only #[update] method with 1 argument (has &self in its arguments).

Encoding is the same as Self::Init1

CallStatefulRo2 = 21

Stateful read-only #[update] method with 2 arguments (has &self in its arguments).

Encoding is the same as Self::Init1

CallStatefulRo3 = 22

Stateful read-only #[update] method with 3 arguments (has &self in its arguments).

Encoding is the same as Self::Init1

CallStatefulRo4 = 23

Stateful read-only #[update] method with 4 arguments (has &self in its arguments).

Encoding is the same as Self::Init1

CallStatefulRo5 = 24

Stateful read-only #[update] method with 5 arguments (has &self in its arguments).

Encoding is the same as Self::Init1

CallStatefulRo6 = 25

Stateful read-only #[update] method with 6 arguments (has &self in its arguments).

Encoding is the same as Self::Init1

CallStatefulRo7 = 26

Stateful read-only #[update] method with 7 arguments (has &self in its arguments).

Encoding is the same as Self::Init1

CallStatefulRo8 = 27

Stateful read-only #[update] method with 8 arguments (has &self in its arguments).

Encoding is the same as Self::Init1

CallStatefulRo9 = 28

Stateful read-only #[update] method with 9 arguments (has &self in its arguments).

Encoding is the same as Self::Init1

CallStatefulRo10 = 29

Stateful read-only #[update] method with 10 arguments (has &self in its arguments).

Encoding is the same as Self::Init1

CallStatefulRw1 = 30

Stateful read-write #[update] method with 1 argument (has &mut self in its arguments).

Encoding is the same as Self::Init1

CallStatefulRw2 = 31

Stateful read-write #[update] method with 2 arguments (has &mut self in its arguments).

Encoding is the same as Self::Init1

CallStatefulRw3 = 32

Stateful read-write #[update] method with 3 arguments (has &mut self in its arguments).

Encoding is the same as Self::Init1

CallStatefulRw4 = 33

Stateful read-write #[update] method with 4 arguments (has &mut self in its arguments).

Encoding is the same as Self::Init1

CallStatefulRw5 = 34

Stateful read-write #[update] method with 5 arguments (has &mut self in its arguments).

Encoding is the same as Self::Init1

CallStatefulRw6 = 35

Stateful read-write #[update] method with 6 arguments (has &mut self in its arguments).

Encoding is the same as Self::Init1

CallStatefulRw7 = 36

Stateful read-write #[update] method with 7 arguments (has &mut self in its arguments).

Encoding is the same as Self::Init1

CallStatefulRw8 = 37

Stateful read-write #[update] method with 8 arguments (has &mut self in its arguments).

Encoding is the same as Self::Init1

CallStatefulRw9 = 38

Stateful read-write #[update] method with 9 arguments (has &mut self in its arguments).

Encoding is the same as Self::Init1

CallStatefulRw10 = 39

Stateful read-write #[update] method with 10 arguments (has &mut self in its arguments).

Encoding is the same as Self::Init1

ViewStateless1 = 40

Stateless #[view] method with 1 argument (doesn’t have self in its arguments).

Encoding is the same as Self::Init1

ViewStateless2 = 41

Stateless #[view] method with 2 arguments (doesn’t have self in its arguments).

Encoding is the same as Self::Init1

ViewStateless3 = 42

Stateless #[view] method with 3 arguments (doesn’t have self in its arguments).

Encoding is the same as Self::Init1

ViewStateless4 = 43

Stateless #[view] method with 4 arguments (doesn’t have self in its arguments).

Encoding is the same as Self::Init1

ViewStateless5 = 44

Stateless #[view] method with 5 arguments (doesn’t have self in its arguments).

Encoding is the same as Self::Init1

ViewStateless6 = 45

Stateless #[view] method with 6 arguments (doesn’t have self in its arguments).

Encoding is the same as Self::Init1

ViewStateless7 = 46

Stateless #[view] method with 7 arguments (doesn’t have self in its arguments).

Encoding is the same as Self::Init1

ViewStateless8 = 47

Stateless #[view] method with 8 arguments (doesn’t have self in its arguments).

Encoding is the same as Self::Init1

ViewStateless9 = 48

Stateless #[view] method with 9 arguments (doesn’t have self in its arguments).

Encoding is the same as Self::Init1

ViewStateless10 = 49

Stateless #[view] method with 10 arguments (doesn’t have self in its arguments).

Encoding is the same as Self::Init1

ViewStatefulRo1 = 50

Stateful read-only #[view] method with 1 argument (has &self in its arguments).

Encoding is the same as Self::Init1

ViewStatefulRo2 = 51

Stateful read-only #[view] method with 2 arguments (has &self in its arguments).

Encoding is the same as Self::Init1

ViewStatefulRo3 = 52

Stateful read-only #[view] method with 3 arguments (has &self in its arguments).

Encoding is the same as Self::Init1

ViewStatefulRo4 = 53

Stateful read-only #[view] method with 4 arguments (has &self in its arguments).

Encoding is the same as Self::Init1

ViewStatefulRo5 = 54

Stateful read-only #[view] method with 5 arguments (has &self in its arguments).

Encoding is the same as Self::Init1

ViewStatefulRo6 = 55

Stateful read-only #[view] method with 6 arguments (has &self in its arguments).

Encoding is the same as Self::Init1

ViewStatefulRo7 = 56

Stateful read-only #[view] method with 7 arguments (has &self in its arguments).

Encoding is the same as Self::Init1

ViewStatefulRo8 = 57

Stateful read-only #[view] method with 8 arguments (has &self in its arguments).

Encoding is the same as Self::Init1

ViewStatefulRo9 = 58

Stateful read-only #[view] method with 9 arguments (has &self in its arguments).

Encoding is the same as Self::Init1

ViewStatefulRo10 = 59

Stateful read-only #[view] method with 10 arguments (has &self in its arguments).

Encoding is the same as Self::Init1

EnvRo = 60

Read-only #[env] argument.

Example: #[env] env: &Env,

EnvRw = 61

Read-write #[env] argument.

Example: #[env] env: &mut Env,

SlotWithAddressRo = 62

Read-only #[slot] argument with an address.

Example: #[slot] (from_address, from): (&Address, &MaybeData<Slot>),

SlotWithAddressRw = 63

Read-write #[slot] argument with an address.

Example: #[slot] (from_address, from): (&Address, &mut MaybeData<Slot>),

SlotWithoutAddressRo = 64

Read-only #[slot] argument without an address.

Example: #[slot] from: &MaybeData<Slot>,

SlotWithoutAddressRw = 65

Read-write #[slot] argument without an address.

Example: #[slot] from: &mut MaybeData<Slot>,

Input = 66

#[input] argument.

Example: #[input] balance: &Balance,

Output = 67

#[output] argument.

Example: #[output] out: &mut VariableBytes<1024>,

Result = 68

Explicit #[result] argument or T of Result<T, ContractError> return type or simply return type if it is not fallible.

Example: #[result] result: &mut MaybeData<Balance>,

NOTE: There is always exactly one result in a method.