ab_contracts_common::env
Struct Env

#[non_exhaustive]

#[repr(C)]
pub struct Env { /* private fields */ }

Ephemeral execution environment

Implementations

impl Env

pub fn context<'a>(self: &'a &'a mut Self) -> &'a Address

Context of the execution

pub fn caller<'a>(self: &'a &'a mut Self) -> &'a Address

Caller of this contract

pub fn own_address(&self) -> &Address

Address of this contract itself

pub fn call<Args>( &self, contract: &Address, args: &mut Args, method_context: &MethodContext, ) -> Result<(), ContractError>
where Args: ExternalArgs,

Call a single method at specified address and with specified arguments.

This is a shortcut for Self::prepare_call_method + Self::call_many.

pub fn prepare_call_method<'a, Args>( contract: &'a Address, args: &'a mut Args, method_context: &'a MethodContext, ) -> PreparedMethod<'a>
where Args: ExternalArgs,

Prepare a single method for invocation at specified address and with specified arguments

pub fn call_many<const N: usize>( &self, methods: [PreparedMethod<'_>; N], ) -> [Result<(), ContractError>; N]

Invoke provided methods and wait for results.

Remaining gas will be split equally between all individual invocations.