Crate ab_contracts_common

Modules

env

metadata

method

Structs

Address

Logically the same as u128, but aligned to 8 bytes instead of 16.

Balance

Logically the same as u128, but aligned to 8 bytes instead of 16.

CustomContractErrorCode

ExitCode

ShardIndex

Shard index

Enums

ContractError

Constants

MAX_CODE_SIZE

Max allowed size of the contract code

Traits

Contract

A trait that indicates the struct is a contact.

ContractTrait

A trait that indicates the implementation of a contract trait by a contract.

ContractTraitDefinition

A trait that is implemented for dyn ContractTrait and includes constants related to trait definition.